scripts

+++ clean.sh
@@ -3,5 +3,6 @@
 for f in $(find . -maxdepth 1 -type d)
 do
 	[[ $f == "." ]] && continue
+	[[ $f == ".git" ]] && continue
 	rm -rf $f
 done
